moved placeholder class to proper place
